backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
PyMuPDF
httpx
pydantic-settings
redis
python-jose[cryptography]